adapter/primary/api/api_test.go
package api

import (
	"net/http"
	"net/http/httptest"
	"os"
	"testing"

	"github.com/labstack/echo/v4"
	"github.com/stretchr/testify/assert"
	"github.com/witalok2/test-dev-golang-api/internal/entity"
)

func TestRun(t *testing.T) {
	os.Setenv("GAE_ENV", entity.PRODUCTION)
	defer os.Unsetenv("GAE_ENV")

	e := New()
	e.GET("/", func(c echo.Context) error {
		return c.String(http.StatusOK, "Hello, World!")
	})

	req := httptest.NewRequest(http.MethodGet, "/", nil)
	rec := httptest.NewRecorder()
	e.ServeHTTP(rec, req)

	assert.Equal(t, http.StatusOK, rec.Code)
	assert.Equal(t, "Hello, World!", rec.Body.String())
}
